requirements.txt
# SENAI Courses Advisor - dependencies
requests>=2.28.0
beautifulsoup4>=4.12.0
python-telegram-bot>=20.0
schedule>=1.2.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0
pytz>=2021.3
click>=8.0.0
pytest>=7.0.0

# Optional: for turmas modal scraping if no API is found
playwright>=1.40.0
